+++ packages/kit/src/exports/vite/index.js
@@ -218,21 +218,6 @@ function kit({ svelte_config }) {
 
 			const generated = path.posix.join(kit.outDir, 'generated');
 
-			// This ensures that esm-env is inlined into the server output with the
-			// export conditions resolved correctly through Vite. This prevents adapters
-			// that bundle later on from resolving the export conditions incorrectly
-			// and for example include browser-only code in the server output
-			// because they for example use esbuild.build with `platform: 'browser'`
-			const noExternal = ['esm-env'];
-
-			// Vitest bypasses Vite when loading external modules, so we bundle
-			// when it is detected to keep our virtual modules working.
-			// See https://github.com/sveltejs/kit/pull/9172
-			// and https://vitest.dev/config/#deps-registernodeloader
-			if (process.env.TEST) {
-				noExternal.push('@sveltejs/kit');
-			}
-
 			// dev and preview config can be shared
 			/** @type {import('vite').UserConfig} */
 			const new_config = {
@@ -269,7 +254,23 @@ function kit({ svelte_config }) {
 					]
 				},
 				ssr: {
-					noExternal
+					noExternal: [
+						// This ensures that esm-env is inlined into the server output with the
+						// export conditions resolved correctly through Vite. This prevents adapters
+						// that bundle later on from resolving the export conditions incorrectly
+						// and for example include browser-only code in the server output
+						// because they for example use esbuild.build with `platform: 'browser'`
+						'esm-env',
+						// We need this for two reasons:
+						// 1. Without this, `@sveltejs/kit` imports are kept as-is in the server output,
+						//    and that causes modules and therefore classes like `Redirect` to be imported twice
+						//    under different IDs, which breaks a bunch of stuff because of failing instanceof checks.
+						// 2. Vitest bypasses Vite when loading external modules, so we bundle
+						//    when it is detected to keep our virtual modules working.
+						//    See https://github.com/sveltejs/kit/pull/9172
+						//    and https://vitest.dev/config/#deps-registernodeloader
+						'@sveltejs/kit'
+					]
 				}
 			};
 
@@ -378,17 +379,34 @@ function kit({ svelte_config }) {
 				case '\0__sveltekit/paths':
 					const { assets, base } = svelte_config.kit.paths;
 
+					// use the values defined in `global`, but fall back to hard-coded values
+					// for the sake of things like Vitest which may import this module
+					// outside the context of a page
 					if (browser) {
-						return `export const base = ${s(base)};
-export const assets = ${global}.assets;`;
+						return `export const base = ${global}?.base ?? ${s(base)};
+export const assets = ${global}?.assets ?? ${assets ? s(assets) : 'base'};`;
 					}
 
-					return `export const base = ${s(base)};
+					return `export let base = ${s(base)};
 export let assets = ${assets ? s(assets) : 'base'};
+
+export const relative = ${svelte_config.kit.paths.relative};
+
+const initial = { base, assets };
+
+export function override(paths) {
+	base = paths.base;
+	assets = paths.assets;
+}
+
+export function reset() {
+	base = initial.base;
+	assets = initial.assets;
+}
 
 /** @param {string} path */
 export function set_assets(path) {
-	assets = path;
+	assets = initial.assets = path;
 }`;
 
 				case '\0__sveltekit/environment':