+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sveltejs/kit
 
+## 1.9.1
+
+### Patch Changes
+
+- feat: warn when calling depends(...) with special URI scheme ([#9246](https://github.com/sveltejs/kit/pull/9246))
+
+## 1.9.0
+
+### Minor Changes
+
+- feat: add `paths.relative` option to control interpretation of `paths.assets` and `paths.base` ([#9220](https://github.com/sveltejs/kit/pull/9220))
+
+## 1.8.8
+
+### Patch Changes
+
+- fix: always add `@sveltejs/kit` to `noExternal` for ssr build ([#9242](https://github.com/sveltejs/kit/pull/9242))
+- feat: add `api.methods` and `page.methods` to `builder.routes` ([#9145](https://github.com/sveltejs/kit/pull/9145))
+
 ## 1.8.7
 
 ### Patch Changes